🔁 move `itemStats` to `Status` cmd class
